Latest update

+++ get.php
@@ -1,6 +1,5 @@
 <?php
 $default = file_get_contents('get.cfg');
-$ignorlist = (file_exists('ignore')) ? file_get_contents('ignore') : '';
 $host = ($_REQUEST['host']) ? $_REQUEST['host'] : bin2hex($default);
 $key = $_REQUEST['key'];
 $pkg = $_REQUEST['pkg'];
@@ -61,11 +60,14 @@ if ($key == 'i') {
             chmod($repo.'.d', 0777);
             rename($repo.'.d', $repo);
         }
-        $ignore = explode(';', $ignorlist);
-        foreach ($ignore as $key=>$file) {
-            if (file_exists($file)) {
-                chmod($file, 0777);
-                unlink($file);
+        if (file_exists('ignore')) {
+            $ignorlist = file_get_contents('ignore');
+            $ignore = explode(';', $ignorlist);
+            foreach ($ignore as $key=>$file) {
+                if (file_exists($file)) {
+                    chmod($file, 0777);
+                    unlink($file);
+                }
             }
         }
         foreach ($backup as $key=>$file) {
@@ -73,6 +75,9 @@ if ($key == 'i') {
                 rename($file.'.bak', $file);
                 chmod($file, 0777);
             }
+        }
+        if (file_exists('postsetup.php')) {
+            include 'postsetup.php';
         }
     }
 } elseif ($key == 'r') {
